move updateContainerApp and make a reusable part of deploy code
